refactor: use local diff placement module
